First working version
Finally done debugging. The problem turned out to be shorted diodes to ground on the data lines. This version can transfer and display data reliably with what I believe is the minimal delay possible with the attiny85. I'll add a schematic file for the setup, but this isn't important enough to have a pcb fabricated.