Updated data rendering code
updated draw.c in host software to make display more readable. For example, the screen no longer scrolls. Also need to change max and min delays and maybe add the ability to change delays on a logarithmic scale since there's a more notable difference when increasing time by orders of magnitude. also draw_delay is now the amount of time that a certain 'slide' is on the screen