Added board
added schematic of hardware setup. I'm not planning on making this into a pcb, but it's nice to have the schematic. The schematic and code might differ in terms of which pins are used on the attiny85 (D+ and D- are switched and some shift register pins might be switched around).